+++ primap2/pm2io/__init__.py
@@ -1,7 +1,7 @@
 """Data reading module of the PRIMAP2 climate policy analysis package."""
 
 
-from ._data_reading import read_wide_csv_file_if
+from ._data_reading import read_long_csv_file_if, read_wide_csv_file_if
 from ._interchange_format import (
     from_interchange_format,
     read_interchange_format,
@@ -9,6 +9,7 @@ from ._interchange_format import (
 )
 
 __all__ = [
+    "read_long_csv_file_if",
     "read_wide_csv_file_if",
     "from_interchange_format",
     "read_interchange_format",